+++ CHANGELOG.md
@@ -10,4 +10,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Changed:
 
 * dependencies have been updated
-
+* if zero or negative capacity is specified, default to `5` and print error (issue #1)